move async msg to models + doc fix
