+++ massa-async-pool/src/lib.rs
@@ -87,16 +87,10 @@
 
 mod changes;
 mod config;
-mod mapping_grpc;
-mod message;
 mod pool;
 
 pub use changes::{AsyncPoolChanges, AsyncPoolChangesDeserializer, AsyncPoolChangesSerializer};
 pub use config::AsyncPoolConfig;
-pub use message::{
-    AsyncMessage, AsyncMessageDeserializer, AsyncMessageInfo, AsyncMessageSerializer,
-    AsyncMessageTrigger, AsyncMessageTriggerSerializer, AsyncMessageUpdate,
-};
 pub use pool::{AsyncPool, AsyncPoolDeserializer, AsyncPoolSerializer};
 
 #[cfg(test)]
