+++ massa-execution-worker/src/active_history.rs
@@ -1,6 +1,6 @@
-use massa_async_pool::{AsyncMessage, AsyncMessageUpdate};
 use massa_execution_exports::ExecutionOutput;
 use massa_ledger_exports::{LedgerEntry, LedgerEntryUpdate};
+use massa_models::async_msg::{AsyncMessage, AsyncMessageUpdate};
 use massa_models::async_msg_id::AsyncMessageId;
 use massa_models::denunciation::DenunciationIndex;
 use massa_models::prehash::{CapacityAllocator, PreHashMap, PreHashSet};
