+++ massa-models/src/lib.rs
@@ -13,6 +13,8 @@ pub mod address;
 /// amount related structures
 pub mod amount;
 /// async message related structures
+pub mod async_msg;
+/// async message id
 pub mod async_msg_id;
 /// block structure
 pub mod block;
